requirements.txt
# The library requirements will be specified here

fastapi~=0.115.4
SQLAlchemy~=2.0.36
passlib~=1.7.4
uvicorn~=0.32.0
pyjwt~=2.10.1
apscheduler~=3.11.0
python-telegram-bot~=22.0
pydantic~=2.11.3
redis~=6.0.0
rq~=2.3.0